ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "mlflow>=3.7.0",
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime-gpu>=1.23.2",
    "onnxscript>=0.5.6",
    "pillow>=12.0.0",
    "psutil>=7.2.1",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.2",
    "torch>=2.9.1",
    "tqdm>=4.67.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=8.0.0",
    "ruff>=0.14.8",
    "tox>=4.32.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
]

[project.scripts]
train = "scripts.train:main"
simple-infer = "scripts.simple_infer:main"
onnx-export = "scripts.export_to_onnx:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["config", "models", "dataset", "training", "inference", "utils", "scripts", "tests",  "../shared"]
exclude = [".venv*", ".tox*", "mlruns*", "__pycache__*", "*.egg-info"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "N", "I", "B", "UP", "TCH", "ANN", "PT"]
ignore = [
    "E501",     # covered by formatter / line length managed elsewhere
    "ANN401",   # allow Any in some cases
    "N806",     # ML is full of upper case names
    "N812",     # ML import is full of upper case names
    "N803",     # I just like to have UPPERCASE sometimes
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
files = [
    "dataset",
    "models",
    "config",
    "training",
    "inference",
    "scripts",
]
plugins = [
    "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[tool.pyright]
include = ["config", "models", "dataset", "training", "inference", "utils", "scripts", "tests","../shared"]
extraPaths = ["."]
pythonVersion = "3.13"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
useLibraryCodeForTypes = false

[[tool.pyright.overrides]]
module = "torch.amp"
reportAttributeAccessIssue = false
reportPrivateImportUsage = false

[[tool.pyright.overrides]]
module = "torch.multiprocessing"
reportAttributeAccessIssue = false

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true
